Create utils::io::net for networking helpers
