+++ lib/src/imdb/service.rs
@@ -1,7 +1,7 @@
 #![warn(clippy::all)]
 
 use std::fs::{self, File};
-use std::io::{BufRead, BufReader, BufWriter};
+use std::io::BufWriter;
 use std::path::Path;
 use std::time::{Duration, Instant};
 
@@ -10,15 +10,13 @@ use crate::imdb::db_binary::ServiceDbFromBinary;
 use crate::imdb::title::Title;
 use crate::imdb::title_id::TitleId;
 use crate::imdb::tsv_import::tsv_import;
-use crate::utils;
-use crate::utils::io::progress::ProgressPipe;
+use crate::utils::io::file as io_file;
+use crate::utils::io::net as io_net;
 use crate::utils::result::Res;
 use crate::utils::search::SearchString;
 
-use flate2::bufread::GzDecoder;
 use humantime::format_duration;
 use log::{debug, log_enabled};
-use reqwest::blocking::{Client, Response};
 use reqwest::Url;
 
 /// Struct providing the movies and series databases and the related services
@@ -68,32 +66,6 @@ impl Service {
     Ok(service)
   }
 
-  /// Sends a GET request to the given URL and returns the response.
-  ///
-  /// # Arguments
-  ///
-  /// * `imdb_url` - The base URL to send the GET request to.
-  /// * `path` - Endpoint path.
-  fn get_response(imdb_url: &Url, path: &str) -> Res<Response> {
-    let url = imdb_url.join(path)?;
-    let client = Client::builder().build()?;
-    let resp = client.get(url).send()?;
-    Ok(resp)
-  }
-
-  /// Returns a reader for the given response.
-  ///
-  /// # Arguments
-  ///
-  /// * `resp` - Response returned for the GET request.
-  /// * `progress_fn` - Function to keep track of the download progress.
-  fn create_fetcher(resp: Response, progress_fn: impl Fn(u64)) -> impl BufRead {
-    let progress = ProgressPipe::new(resp, progress_fn);
-    let reader = BufReader::new(progress);
-    let decoder = GzDecoder::new(reader);
-    BufReader::new(decoder)
-  }
-
   /// Ensures that the movies and series databases exist and are up-to-date.
   ///
   /// The databases are created if they don't exist, and updated if they are outdated or
@@ -114,8 +86,8 @@ impl Service {
   ) -> Res {
     let needs_update = {
       force_db_update
-        || utils::io::file::file_older_than(&utils::io::file::file_exists(movies_db_filename)?, max_age)
-        || utils::io::file::file_older_than(&utils::io::file::file_exists(series_db_filename)?, max_age)
+        || io_file::file_older_than(&io_file::file_exists(movies_db_filename)?, max_age)
+        || io_file::file_older_than(&io_file::file_exists(series_db_filename)?, max_age)
     };
 
     if needs_update {
@@ -131,8 +103,8 @@ impl Service {
       let series_db_writer = BufWriter::new(series_db_file);
 
       let imdb_url = Url::parse(IMDB_URL)?;
-      let basics_response = Self::get_response(&imdb_url, BASICS_FILENAME)?;
-      let ratings_response = Self::get_response(&imdb_url, RATINGS_FILENAME)?;
+      let basics_response = io_net::get_response(imdb_url.join(BASICS_FILENAME)?)?;
+      let ratings_response = io_net::get_response(imdb_url.join(RATINGS_FILENAME)?)?;
 
       let content_length = match (basics_response.content_length(), ratings_response.content_length()) {
         (None, _) | (_, None) => None,
@@ -143,8 +115,8 @@ impl Service {
 
       progress_fn(content_length, 0);
 
-      let basics_fetcher = Self::create_fetcher(basics_response, |bytes| progress_fn(None, bytes));
-      let ratings_fetcher = Self::create_fetcher(ratings_response, |bytes| progress_fn(None, bytes));
+      let basics_fetcher = io_net::create_fetcher(basics_response, |bytes| progress_fn(None, bytes));
+      let ratings_fetcher = io_net::create_fetcher(ratings_response, |bytes| progress_fn(None, bytes));
 
       tsv_import(ratings_fetcher, basics_fetcher, movies_db_writer, series_db_writer)?;
     } else {
