+++ lib/src/utils/io/mod.rs
@@ -4,4 +4,5 @@
 //! Common utilities for things like networking and IO.
 
 pub mod file;
+pub mod net;
 pub mod progress;
